weather_app_backend/requirements.txt
Django
djangorestframework
requests
django-cors-headers
python-decouple
pytz
mongoengine
pymongo
djongo
openpyxl
requests-cache 
retry-requests 
numpy 
pandas
pywttr 
pywttr-models
aiopywttr